server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "motor>=3.7.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.3",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "slowapi>=0.1.9",
]
